feat(logger): introduce centralized logger service
Replace scattered logging implementations with a unified LoggerService
